pyproject.toml
[project]
name = "django-react-session-auth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.3",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.16.0",
    "drf-spectacular[sidecar]>=0.28.0",
]
